added snuke.cpp
Problem Statement

There is a grid with H horizontal rows and W vertical columns. Each cell has a lowercase English letter written on it. We denote by (i,j) the cell at the i-th row from the top and j-th column from the left.

The letters written on the grid are represented by H strings S1,S2,…,SH, each of length W. The j-th letter of Si represents the letter written on (i,j).

There is a unique set of contiguous cells (going vertically, horizontally, or diagonally) in the grid with s, n, u, k, and e written on them in this order.
Find the positions of such cells and print them in the format specified in the Output section.

A tuple of five cells (A1,A2,A3,A4,A5) is said to form a set of contiguous cells (going vertically, horizontally, or diagonally) with s, n, u, k, and e written on them in this order if and only if all of the following conditions are satisfied.

    A1,A2,A3,A4 and A5 have letters s, n, u, k, and e written on them, respectively.
    For all 1≤i≤4, cells Ai and Ai+1 share a corner or a side.
    The centers of A1,A2,A3,A4, and A5 are on a common line at regular intervals.

Constraints

    5≤H≤100
    5≤W≤100
    H and W are integers.
    Si is a string of length W consisting of lowercase English letters.
    The given grid has a unique conforming set of cells.

Input

The input is given from Standard Input in the following format:

H W
S1
S2
⋮
SH

Output

Print five lines in the following format.

Let (R1,C1),(R2,C2)…,(R5,C5) be the cells in the sought set with s, n, u, k, and e written on them, respectively. The i-th line should contain Ri and Ci in this order, separated by a space.

In other words, print them in the following format:

R1 C1
R2 C2
⋮
R5 C5
[B

I used the code for the findwords assignment and figured out that I can use it again for this problem with a little bit of adjustment because why not